==> Step 1 - Create corpus.ipynb <==
### Step 1 - Create corpus
Creates a gzipped DataFrame from a corpus of raw text files.

from nltk.corpus.reader import CategorizedPlaintextCorpusReader
from pandas import DataFrame
import numpy as np

corpus = CategorizedPlaintextCorpusReader("./corpus/lyrics/raw/", r".*\.txt", cat_pattern=r"(\w+)/*")

lyrics = []
for filen in corpus.fileids():
    with open("./corpus/lyrics/raw/" + filen, 'rb') as handle:
        lyrics.append(handle.read())
        
arr = np.array([corpus.fileids(), lyrics])
lyrics_df = DataFrame(arr.transpose(), columns=['uri', 'lyrics'])
lyrics_df.head()

lyrics_df.info()

lyrics_df.to_csv("./corpus/lyrics/lyrics_raw.gz", compression="gzip", header=None)
