==> Step 2 - Tokenize.ipynb <==
## Step 2 - Tokenize

from __future__ import division
from nltk.corpus.reader import CategorizedPlaintextCorpusReader
import itertools
import nltk
import string
import re
import pandas as pd
from pandas import Series

### Load data and create a Series
Load a compressed dataframe and converts to a Series after a first pass at cleaning up the text.

def initialClean(text):
    # remove newlines, doubled singlequotes, excess whitespace
    text = text.replace("\n", " ").replace("\'\'", "").strip()
    # remove text enclosed in brackets; ex: [Verse 1]
    text = re.sub(r'\[.+?\]', '', text).strip()
    # remove text beginning with a single quote; ex: '01, 'em
    text = re.sub(r"\'\w+", "", text).strip()
    return text

# read the raw lyrics from a Pandas DataFrame saved as a csv.
lyrics_df = pd.read_csv("./corpus/lyrics/lyrics_raw.gz", compression='gzip', names=['uri', 'lyrics'])
lyrics_df.head()

lyrics_raw = []
for lyrics in lyrics_df['lyrics']:
    lyrics_raw.append(initialClean(lyrics))

# create a Series to index lyrics by filename
lyrics_by_file = Series(lyrics_raw,index=lyrics_df['uri'])

### Tokenize
We'll tokenize using a regular expression tokenizer to preserve contractions.

from nltk.tokenize import RegexpTokenizer
regexpTokenizer = RegexpTokenizer("[\w']+")

for filen, lyrics in lyrics_by_file.iteritems():
    lyrics_by_file[filen] = " ".join(regexpTokenizer.tokenize(lyrics))
lyrics_by_file.head()

from nltk.corpus import stopwords
english_stops = stopwords.words('english')
custom_stops = []
with open("./resources/custom_stopwords.txt", 'rb') as handle:
    custom_stops = handle.read().split("\n")
    print custom_stops

def clean_tokens(tokens):
    """ Lowercases, takes out punct and stopwords and short strings """
    return [token.lower() for token in tokens if (token not in string.punctuation) and 
                   (token.lower() not in custom_stops) and (token.lower() not in english_stops) and len(token) > 2]

lyrics_clean = []
for filen, lyrics in lyrics_by_file.iteritems():
    lyrics_clean.append(" ".join(clean_tokens(lyrics.split(' '))))
    
# create a Series to index cleaned lyrics by filename
lyrics_clean_by_file = Series(lyrics_clean,index=lyrics_by_file.index)
lyrics_clean_by_file.head()

total_raw_tokens = 0
for filen, lyrics in lyrics_by_file.iteritems():
    total_raw_tokens += len(lyrics.split(' '))

total_clean_tokens = 0
lyrics_clean = ""
for filen, lyrics in lyrics_clean_by_file.iteritems():
    total_clean_tokens += len(lyrics.split(' '))
    lyrics_clean = lyrics_clean + lyrics
    
print "total raw tokens: {0}".format(total_raw_tokens)
print "total clean tokens: {0}".format(total_clean_tokens)
print "{0:.0f}% reduction".format(100 - total_clean_tokens / total_raw_tokens * 100)

# combine all cleaned tokens in an list an sort
%time sorted_clean = sorted(lyrics_clean.split(' '))

# sample clean tokens to see if we still have garbage
sorted_clean[0:10]

### Save tokens
We'll use this file in later notebooks.

def writeFile(path, text):
    text_file = open(path, "w") 
    text_file.write(text + " ")
    text_file.close()

for filen, lyrics in lyrics_clean_by_file.iteritems():
    writeFile("./corpus/lyrics/tokenized/" + filen.replace("spotify:track:", "") + ".txt", lyrics)
