==> Step 3 - Tf-idf create.ipynb <==
### Step 2 - Tf-idf create
Create a file with tfidf scores for each word in the corpus.

WARNING: This process can take several hours.

from __future__ import division
from nltk.corpus.reader import CategorizedPlaintextCorpusReader
import pandas as pd
from pandas import Series
from nltk.text import TextCollection

corpus = CategorizedPlaintextCorpusReader("./corpus/lyrics/tokenized/", r".*\.txt", cat_pattern=r"(\w+)/*")

lyrics = []
for filen in corpus.fileids():
    with open("./corpus/lyrics/tokenized/" + filen, 'rb') as handle:
        lyrics.append(handle.read())

import collections
from nltk import Text

def compute_tfidf_by_doc(coll, texts, filenames):
    tfidf_by_doc = collections.defaultdict(list)
    for i, text in enumerate(texts):
        for word in set(text.split(" ")):   # just use the words in this text
            tfidfscore = coll.tf_idf(word, text)
            tf = coll.tf(word, text) # is actually count / len(text)
            count = text.count(word)
            if tfidfscore:
                tfidf_by_doc[filenames[i]].append({
                    "word": word,
                    "tfidf": tfidfscore
                    #"tf": tf,
                    #"count": count
                })
    return tfidf_by_doc

# We need to make the text collection, then use it to compute the tf-idf for the words in the docs.
collection = TextCollection(corpus)

# Use the first set of lyrics to test out idf and tf_idf scores.
# low tf_idf score - A word like 'yeah' should appear frequently in all songs.
# high tf_idf score - A word like 'feel' may not appear frequently in all songs but occur a lot in a specific song.
lyrics[0]

collection.tf_idf("love", lyrics[0])

collection.tf_idf("know", lyrics[0])

collection.tf_idf("thanks", lyrics[0])

# creates the following dictionary w filename as the key and list of word, counts, tf and tfidf scores. 
# defaultdict(list,
    # {'04CpzA2BdOLEz7EMp5uwTU.txt': [{'count': 1,
        # 'tf': 0.006211180124223602,
        # 'tfidf': 0.01568775555470966,
        # 'word': 'summer'},
        # {'count': 1,
        # 'tf': 0.006211180124223602,
        # 'tfidf': 0.01999301754576522,
        # 'word': 'nights'},
%time tfidfs = compute_tfidf_by_doc(collection, lyrics, corpus.fileids())

len(tfidfs)

### Save tfidf json

import json
jsonified = json.dumps(tfidfs)
with open('./corpus/lyrics_tfidf.json', 'w') as handle:
    handle.write(jsonified)
