==> Step 5 - Frequency distribution.ipynb <==
### Frequency distribution
Word counts and frequency distributions by corpus.

from nltk import Text
from nltk.corpus.reader import CategorizedPlaintextCorpusReader

tokenized_corpus = CategorizedPlaintextCorpusReader("./corpus/lyrics/tokenized/", r".*\.txt", cat_pattern=r"(\w+)/*")
tfidf_corpus = CategorizedPlaintextCorpusReader("./corpus/lyrics/tfidf/", r".*\.txt", cat_pattern=r"(\w+)/*")

if (len(tokenized_corpus.fileids()) > 0):
    print "Tokenized corpus total words: {0}".format(len(tokenized_corpus.words()))
    tokenized_text = Text(tokenized_corpus.words())
    fdist_tokenized = tokenized_text.vocab().most_common()
if (len(tfidf_corpus.fileids()) > 0):
    print "Tfidf corpus total words: {0}".format(len(tfidf_corpus.words()))
    tfidf_text = Text(tfidf_corpus.words())
    fdist_tfidf = tfidf_text.vocab().most_common()

fdist_tokenized[:10]

fdist_tfidf[:10]
