==> Step 6 - Stems.ipynb <==
### Step 4 - Stemming
Reduce the cleaned tokens to their stems. We'll later get the frequency distribution of the stems.

from __future__ import division
from nltk.corpus.reader import CategorizedPlaintextCorpusReader
import nltk
import re
import pandas as pd
from pandas import Series
from nltk.stem.snowball import SnowballStemmer
stemmer = SnowballStemmer("english")
outputFile = "lyrics_stems.txt"

def stem(tokens):
    filtered_tokens = []
    # filter out any tokens not containing letters (e.g., numeric tokens, raw punctuation)
    for token in tokens:
        tk = token.decode('utf8')
        if re.search('[a-zA-Z]', tk):
            filtered_tokens.append(tk)
    stems = [stemmer.stem(t) for t in filtered_tokens]
    return stems

corpus = CategorizedPlaintextCorpusReader("./corpus/lyrics/tfidf/", r".*\.txt", cat_pattern=r"(\w+)/*", encoding="utf8")

lyrics = []
for filen in corpus.fileids():
    with open("./corpus/lyrics/tokenized/" + filen, 'rb') as handle:
        lyrics.append(handle.read())

# create a Series to index lyrics by filename
lyrics_by_file = Series(lyrics,index=corpus.fileids())
lyrics_by_file.head()

### Stem
We'll stem using the Snowball stemmer.

lyrics_stemmed = []
for filen, lyrics in lyrics_by_file.iteritems():
    lyrics_stemmed.append(' '.join(stem(lyrics.split(' '))))

# create a Series to index cleaned lyrics by filename
lyrics_stemmed_by_file = Series(lyrics_stemmed,index=corpus.fileids())
lyrics_stemmed_by_file.head()

def writeFile(path, text):
    text_file = open(path, "w") 
    text_file.write(text + " ")
    text_file.close()

for filen, lyrics in lyrics_stemmed_by_file.iteritems():
    writeFile("./corpus/lyrics/stemmed/" + filen, lyrics.encode("utf8"))
